calculate the lightning distance to us
"One of my friends was a mountaineer, and he used to say that we had to figure out how far lightning was from us when it was raining from a distance, so we could take cover. He told me I could write a program to calculate the lightning distance to us. I wanted to help my friend, so I'm sharing this code with you. Good luck!"